pyproject.toml
[tool.poetry]
name = "people_control"
version = "0.1.0"
description = "Um pacote rápido de depêndencias para um projeto com django e vuejs"
authors = ["Ronald Seabra <git.ronaldseabra@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
python-dotenv = "^1.0.0"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.269"
black = "^23.3.0"
model-bakery = "^1.11.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
extend-exclude = ["core/migrations"]

[tool.ruff]
line-length = 120
extend-exclude = ["core/migrations"]
target-version = "py311"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "people_control.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py","*_tests.py", "testing/python/*.py"]
